backend/requirements.txt
faster-whisper==1.1.0
openai==1.60.1
flask==3.1.0
flask-cors==5.0.0
python-dotenv==1.0.1
moviepy==1.0.3